docs(readme): remove outdated deployed app link from usage instructions
The previous README included a link to a deployed version of the app that is no longer available or maintained. This commit cleans up the usage section by removing the broken or outdated URL, leaving only instructions to open the local `index.html` file. This helps avoid confusion for new users.